Replace `fromIntegral` with `fromInteger` or `toInteger` where possible.
See #637.

+++ src/Cryptol/REPL/Monad.hs
@@ -793,7 +793,7 @@ userOptions  = mkOptionMap
     "Enable type-checker debugging output." $
     \case EnvNum n -> do me <- getModuleEnv
                          let cfg = M.meSolverConfig me
-                         setModuleEnv me { M.meSolverConfig = cfg{ T.solverVerbose = fromIntegral n } }
+                         setModuleEnv me { M.meSolverConfig = cfg{ T.solverVerbose = n } }
           _        -> return ()
   , OptionDescr "core-lint" (EnvBool False)
     noCheck
